move FileView.AdaptiveLayoutButton to Buttons/ShrinkingIconButton
