+++ catalog/app/components/Buttons/index.ts
@@ -1,0 +1,1 @@
+export { default as ShrinkingIconButton } from './ShrinkingIconButton'
