+++ catalog/app/containers/Bucket/FileView.js
@@ -4,6 +4,7 @@ import * as redux from 'react-redux'
 import * as M from '@material-ui/core'
 
 // import Message from 'components/Message'
+import ShrinkingIconButton from 'components/Buttons/ShrinkingIconButton'
 import SelectDropdown from 'components/SelectDropdown'
 import cfg from 'constants/config'
 import { tokens as tokensSelector } from 'containers/Auth/selectors'
@@ -13,7 +14,19 @@ export * from './Meta'
 
 // TODO: move here everything that's reused btw Bucket/File, Bucket/PackageTree and Embed/File
 
-const useAdaptiveButtonStyles = M.makeStyles((t) => ({
+export function DownloadButton({ className, handle }) {
+  return AWS.Signer.withDownloadUrl(handle, (url) => (
+    <ShrinkingIconButton
+      className={className}
+      href={url}
+      download
+      label="Download file"
+      icon="arrow_downward"
+    />
+  ))
+}
+
+const useViewModeSelectorStyles = M.makeStyles((t) => ({
   root: {
     flexShrink: 0,
     marginBottom: -3,
@@ -24,47 +37,8 @@ const useAdaptiveButtonStyles = M.makeStyles((t) => ({
   },
 }))
 
-export function AdaptiveButtonLayout({ className, label, icon, ...props }) {
-  const classes = useAdaptiveButtonStyles()
-  const t = M.useTheme()
-  const sm = M.useMediaQuery(t.breakpoints.down('sm'))
-
-  return sm ? (
-    <M.IconButton
-      className={cx(classes.root, className)}
-      edge="end"
-      size="small"
-      {...props}
-    >
-      <M.Icon>{icon}</M.Icon>
-    </M.IconButton>
-  ) : (
-    <M.Button
-      className={cx(classes.root, className)}
-      variant="outlined"
-      size="small"
-      startIcon={<M.Icon>{icon}</M.Icon>}
-      {...props}
-    >
-      {label}
-    </M.Button>
-  )
-}
-
-export function DownloadButton({ className, handle }) {
-  return AWS.Signer.withDownloadUrl(handle, (url) => (
-    <AdaptiveButtonLayout
-      className={className}
-      href={url}
-      download
-      label="Download file"
-      icon="arrow_downward"
-    />
-  ))
-}
-
 export function ViewModeSelector({ className, ...props }) {
-  const classes = useAdaptiveButtonStyles()
+  const classes = useViewModeSelectorStyles()
   const t = M.useTheme()
   const sm = M.useMediaQuery(t.breakpoints.down('sm'))
   return (
@@ -86,7 +60,7 @@ export function ZipDownloadForm({ className, suffix, label, newTab = false }) {
       style={{ flexShrink: 0 }}
     >
       <input type="hidden" name="token" value={token} />
-      <AdaptiveButtonLayout
+      <ShrinkingIconButton
         className={className}
         label={label}
         icon="archive"
@@ -118,7 +92,7 @@ const renderDownload = (handle) => !!handle && <DownloadButton {...{ handle }} /
 
 function FileView({
   header,
-  subheader, 
+  subheader,
 }) {
   return (
     <Root>
